Fix tests failing because of bug fix

+++ cheetah/particles.py
@@ -905,7 +905,7 @@ class ParticleBeam(Beam):
         particle_charges = torch.from_numpy(particle_charges)
         return cls(
             particles=particles_7d,
-            energy=torch.tensor(energy),
+            energy=torch.tensor(energy, dtype=torch.float32),
             particle_charges=particle_charges,
             **kwargs,
         )
